pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "fastmcp>=2.14.4",
    "google-genai>=1.61.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "trafilatura>=2.0.0",
    "uvicorn>=0.40.0",
]
